refac: add `ReadOnlyBuffer`
