+++ tachyon/math/finite_fields/cubic_extension_field.h
@@ -382,7 +382,7 @@ class Copyable<Derived, std::enable_if_t<std::is_base_of_v<
   }
 
   static bool ReadFrom(
-      const Buffer& buffer,
+      const ReadOnlyBuffer& buffer,
       math::CubicExtensionField<Derived>* cubic_extension_field) {
     typename Derived::BaseField c0;
     typename Derived::BaseField c1;
